Update
Lots of player controller and prefab stuff.  Broke most of the Player Controller stuff down into a parent Pawn Controller class because it can be used by any pawn.  To further facilitate this, broke down methods that were doing more than one thing into constituent methods.  Now that the Player Controller was bare and inheriting all the mechanics, moved the input controller into it.

Next I want to tackle a Wall script for walls, first feature fading the wall when the player is behind it.