Update
Some multiplayer and movement tweaks.  Code structure reworked a bit to tidy up adhoc additions from the last couple of days.

Changed design for walls from blocks that take up an entire grid space to panels that are on the spaces between grids.

Trying to wrap my head around handling NPCs.  Specifically, pathfinding to the player.  The BFS implementation entirely excludes occupied tiles so in order to get to the tile next to the player I'm considering generating a path backwards from the player that is allowed to ignore occupied tiles, and then generating a path to the second element of the first path (ie. the first element is the player's tile, the second element is the tile next to the player) and moving along it.  I'm thinking there needs to be a smoother way -- just need a bit of clarity.