Update
Properly network spawning enemy prefab so it syncs on all clients.

Added sight range for enemy, it targets the first player to come into range and untargets/changes target if they move out of range.  Super primitive, but it's the next step in getting the enemy to approach the player.

Toying with the idea of generating a path to each of the walkable tile neighbors of the target and selecting the closest (or first in order if a tie).  If no walkable tile neighbors, change target?  Use range if available?

It seems that there is still some build error that causes pathfinding to glitch on build only -- but not for all builds.  Deleting the whole build and re-building usually fixes it.  I'm not sure what's happening with that yet.