Removed Network Code
Removed all of the multiplayer code for UNET.  Unity Networking cannot handle the game as envisioned.  It was initially used so that decisions would be made from a multiplayer mindset, but rapidly as new ideas are added it was becoming intelgral to the code and I would rather end up with something that functions in single player and has to be adapted for multiplayer than something that takes heavy modification to remove UNET later, and then build in a networking system that can handle the game.

Also added a Sprint speed for pawns.