Colliders, Inventory, Equipment
Disabled collision on all colliders (they are now triggers for RayCasts and detection only).  Disabled all RigidBodies All movement in CodeIso is mathematical and grid based.  There is no free motion.  For the time being, that means that gravity isn't needed.  Characters should never "fall over" or need to by affected by Physics.

Added LayerMask to the RayCast that detects what tile a Pawn is standing on.  This fixed a bug where pawns would get stuck if they moved onto a tile with an object (ie. an item) that does not obstruct movement.  They were detecting the object as their current tile.

Tidied up the Inventory system and it's in a good working place now.  UI to interact with it is a little lacking at the moment.

Build Equipped system and did initial integration with Inventoy.  Weapons and Armor properly equip, unquip, and replace one another in the correct slots.  Next will be actually modifyting stats based on equipment.  And then maybe some UI to work with it.

I want to rebuild the movement system, but I think I will wait until implementing tiles that modify movement speed because weighted pathfinding will be needed then.