Inventory UI Work
Fleshed out most of the Inventory UI.  It not appears and disappears via hotkey, initializes and sizes itself (and can be updated on the fly) based on capacity (ie. what bag you use) and shows the icon of the item that occupies each space.  Also preserves empty spaces!

Next will be some mouse interaction (and the RayCast changes needed... might result in reworking all of the interactable stuff into it's own separate script) and a character/equipment screen.