Fixed Equipment Array
Unified the Armor and Weapons array into one Items array that can be indexed directly against the equipped slots and reworked the logic to work in Items rather than subclasses.

Rethinking how inventory slots are updated.  Might move the methods to the slots themselves.  This would allow the slot to call an overridable Use method on the items themselves for maximum adaptability and scalability.