Drag & Drop
Added drag & drop to the Inventory to move items around from slot to slot.

Need to implement interaction with the Equipment slots....  which feels like it's going to be a pain.  I might investigate re-building this system so that both types of slot use the same Class and identify type with an Enum.

Also want to allow people to drag things and drop them on the ground in their tile, an adjacent tile, or (ultimately) on another player to initial trade.